xadmin-server/requirements.txt
django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
mysqlclient==2.2.4
django-redis==5.4.0
pycryptodomex~=3.21.0
djangorestframework-simplejwt==5.3.1
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
flower==2.0.1
python-daemon==3.0.2
gunicorn==23.0.0
django-proxy==1.3.0
psutil==5.9.8
uvicorn==0.30.6
daphne==4.1.2
channels==4.1.0
channels-redis==4.2.0
django-ranged-response==0.2.0
user-agents==2.2.0
aiofiles==24.1.0
websockets==13.0.1
django-imagekit==5.0.0
pilkit==3.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
openpyxl~=3.1.5
pyzipper==0.3.6
unicodecsv==0.14.1
chardet==5.2.0
pyexcel==0.7.0
pyexcel-xlsx==0.6.0
alibabacloud-dysmsapi20170525==3.0.0
phonenumbers==8.13.44
pycountry==24.6.1
geoip2==4.8.0
ipip-ipdb==1.6.1
requests==2.32.3
html2text==2024.2.26
pandas~=2.2.3
six~=1.16.0
alibabacloud_tea_openapi~=0.3.12
redis~=5.1.1
pillow~=10.4.0
asgiref~=3.8.1